Include stand alone Float types into default ObjectMapper
