+++ ForgeUnity/Assets/BeardedManStudios/Editor/Resources/BMS_Forge_Editor/StandAloneNetworkObjectTemplate.txt
@@ -1,6 +1,5 @@
 using BeardedManStudios.Forge.Networking.Frame;
 using BeardedManStudios.Forge.Networking.Unity;
-using BeardedManStudios.Forge.Networking.StandAlone;
 using System;
 
 namespace BeardedManStudios.Forge.Networking.Generated
@@ -57,7 +56,7 @@ namespace BeardedManStudios.Forge.Networking.Generated
 			base.OwnershipChanged();
 			SnapInterpolations();
 		}
-		
+
 		public void SnapInterpolations()
 		{
 			>:FOREVERY variables:<
@@ -70,7 +69,7 @@ namespace BeardedManStudios.Forge.Networking.Generated
 		protected override BMSByte WritePayload(BMSByte data)
 		{
 			>:FOREVERY variables:<
-			StandAloneObjectMapper.Instance.MapBytes(data, _>:[1]:<);
+			ObjectMapper.Instance.MapBytes(data, _>:[1]:<);
 			>:ENDFOREVERY:<
 
 			return data;
@@ -79,7 +78,7 @@ namespace BeardedManStudios.Forge.Networking.Generated
 		protected override void ReadPayload(BMSByte payload, ulong timestep)
 		{
 			>:FOREVERY variables:<
-			_>:[1]:< = StandAloneObjectMapper.Instance.Map<>:[0]:<>(payload);
+			_>:[1]:< = ObjectMapper.Instance.Map<>:[0]:<>(payload);
 			>:[1]:<Interpolation.current = _>:[1]:<;
 			>:[1]:<Interpolation.target = _>:[1]:<;
 			RunChange_>:[1]:<(timestep);
@@ -93,7 +92,7 @@ namespace BeardedManStudios.Forge.Networking.Generated
 
 			>:FOREVERY variables:<
 			if ((>:[5]:< & _dirtyFields[>:[6]:<]) != 0)
-				StandAloneObjectMapper.Instance.MapBytes(dirtyFieldsData, _>:[1]:<);
+				ObjectMapper.Instance.MapBytes(dirtyFieldsData, _>:[1]:<);
 			>:ENDFOREVERY:<
 
 			// Reset all the dirty fields
@@ -116,12 +115,12 @@ namespace BeardedManStudios.Forge.Networking.Generated
 			{
 				if (>:[1]:<Interpolation.Enabled)
 				{
-					>:[1]:<Interpolation.target = StandAloneObjectMapper.Instance.Map<>:[0]:<>(data);
+					>:[1]:<Interpolation.target = ObjectMapper.Instance.Map<>:[0]:<>(data);
 					>:[1]:<Interpolation.Timestep = timestep;
 				}
 				else
 				{
-					_>:[1]:< = StandAloneObjectMapper.Instance.Map<>:[0]:<>(data);
+					_>:[1]:< = ObjectMapper.Instance.Map<>:[0]:<>(data);
 					RunChange_>:[1]:<(timestep);
 				}
 			}
